backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pdfplumber==0.11.4
python-multipart==0.0.12
typer==0.12.5
supabase==2.9.0
python-dotenv==1.0.1